requirements.txt
jupyter
notebook
pandas
matplotlib
seaborn
scipy
requests
numpy
scikit-learn
plotly
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pydantic-settings
pyyaml
pytest
python-jose[cryptography]
passlib[bcrypt]
python-multipart
